# Optimise bitmap traversal during dumping