Before:

```
$ time asif dump-only --source ./cloud-address-feed.asif --segment 6 > result2.txt
asif dump-only --source ./cloud-address-feed.asif --segment 6 > result2.txt  10.02s user 0.62s system 99% cpu 10.653 total
$ head result2.txt
==== [6] bitmap/akamai ====
104.64.16.0     (1749028864)
104.64.16.1     (1749028865)
104.64.16.2     (1749028866)
104.64.16.3     (1749028867)
104.64.16.4     (1749028868)
104.64.16.5     (1749028869)
104.64.16.6     (1749028870)
104.64.16.7     (1749028871)
104.64.16.8     (1749028872)
$ tail result2.txt
104.64.16.246   (1749029110)
104.64.16.247   (1749029111)
104.64.16.248   (1749029112)
104.64.16.249   (1749029113)
104.64.16.250   (1749029114)
104.64.16.251   (1749029115)
104.64.16.252   (1749029116)
104.64.16.253   (1749029117)
104.64.16.254   (1749029118)
104.64.16.255   (1749029119)
```

After:

```
$ time asif dump-only --source ./cloud-address-feed.asif --segment 6 > result.txt
asif dump-only --source ./cloud-address-feed.asif --segment 6 > result.txt  13.56s user 2.50s system 99% cpu 16.171 total
$ head result.txt
==== [6] bitmap/akamai ====
104.64.16.0      (1749028864)
104.64.16.1      (1749028865)
104.64.16.2      (1749028866)
104.64.16.3      (1749028867)
104.64.16.4      (1749028868)
104.64.16.5      (1749028869)
104.64.16.6      (1749028870)
104.64.16.7      (1749028871)
104.64.16.8      (1749028872)
$ tail result.txt
104.64.16.246    (1749029110)
104.64.16.247    (1749029111)
104.64.16.248    (1749029112)
104.64.16.249    (1749029113)
104.64.16.250    (1749029114)
104.64.16.251    (1749029115)
104.64.16.252    (1749029116)
104.64.16.253    (1749029117)
104.64.16.254    (1749029118)
104.64.16.255    (1749029119)
```